pyproject.toml
[project]
name = "futdata"
version = "0.1.0"
description = "A project designed for football fans to go through their teams latest data!"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "supabase>=2.28.2",
]
